ai-engine/requirements.txt
absl-py==1.4.0
accelerate==1.3.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anthropic==0.36.1
anyio==4.6.2.post1
appnope==0.1.3
APScheduler==3.6.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
av==10.0.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.6.2
boto3==1.35.42
botocore==1.35.42
bs4==0.0.1
cachetools==4.2.2
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.0
click==8.1.7
cmake==3.26.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cPython==0.0.6
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.6.7
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distro==1.9.0
dnspython==1.16.0
email_validator==2.2.0
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.0
exceptiongroup==1.2.2
executing==1.2.0
faiss-cpu==1.9.0
fastapi==0.115.2
fastapi-cli==0.0.5
fastjsonschema==2.16.3
filelock==3.16.1
filetype==1.2.0
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==23.5.9
fonttools==4.38.0
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.18.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.8.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.2
idna==3.10
ifaddr==0.2.0
ImageHash==4.3.1
importlib-metadata==6.6.0
imutils==0.5.4
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.2.0
jovian==0.2.41
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.13.1rc0
kiwisolver==1.4.4
langchain==0.3.7
langchain-anthropic==0.2.3
langchain-aws==0.2.2
langchain-community==0.3.7
langchain-core==0.3.33
langchain-google-genai==2.0.9
langchain-groq==0.2.4
langchain-huggingface==0.1.2
langchain-mongodb==0.4.0
langchain-openai==0.2.8
langchain-postgres==0.0.12
langchain-text-splitters==0.3.0
langgraph==0.2.50
langgraph-checkpoint==2.0.4
langgraph-sdk==0.1.36
langsmith==0.1.135
libclang==16.0.0
lxml==5.3.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multimethod==1.9.1
multiprocess==0.70.14
multitasking==0.0.11
mypy-extensions==1.0.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.54.4
opencv-python==4.7.0.72
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pandas-profiling==3.6.6
parso==0.8.3
patsy==0.5.3
peewee==3.17.8
pexpect==4.8.0
pgvector==0.2.5
phik==0.12.3
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==4.3.6
plotly==5.14.1
plotly-express==0.4.1
prompt-toolkit==3.0.38
propcache==0.2.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.4
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
psycopg2==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.8.1b0
pydot==1.4.2
pyee==9.1.0
Pygments==2.18.0
pylibsrtp==0.8.0
pymongo==4.11
Pympler==1.0.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
PyQt6==6.4.2
PyQt6-Qt6==6.4.3
PyQt6-sip==13.4.1
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
python-pptx==0.6.21
python-telegram-bot==13.11
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywifi==1.1.12
PyYAML==6.0.2
pyzmq==25.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rich==13.9.2
rsa==4.9
s3transfer==0.10.3
safetensors==0.5.2
schedule==1.1.0
scikit-learn==1.2.2
scipy==1.15.1
seaborn==0.12.2
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlmodel==0.0.22
stack-data==0.6.2
starlette==0.40.0
statsmodels==0.13.5
streamlit==1.22.0
streamlit-webrtc==0.45.0
sympy==1.13.1
tangled-up-in-unicode==0.2.0
tavily-python==0.5.0
tenacity==8.5.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow==2.13.0rc0
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc0
termcolor==2.3.0
threadpoolctl==3.1.0
tiktoken==0.8.0
tkcalendar==1.6.1
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.0
torch==2.6.0
tornado==6.2
tqdm==4.66.5
traitlets==5.9.0
transformers==4.48.2
typeguard==2.13.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.20
uuid==1.30
uvicorn==0.32.0
uvloop==0.21.0
validators==0.20.0
visions==0.7.5
watchfiles==0.24.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==13.1
Werkzeug==2.3.4
wrapt==1.14.1
XlsxWriter==3.0.9
xxhash==3.2.0
yarl==1.15.4
ydata-profiling==4.1.2
yfinance==0.2.49
zipp==3.15.0
